Fix class mismatch in velocityinterp track velocity assertion
CI's dedicated ULM Tests job actually has the Curve Fitting Toolbox,
so test_velocityinterp_returns_position_and_velocity_columns ran for
real there (instead of self-skipping like it does locally) and failed:
verifyEqual's AbsTol check still requires matching classes, and
ulm.tracking2D returns tracks as single while the test compared
against double literals. The underlying velocity math was correct
(single 1 vs double 1, single 0 vs double 0) -- cast the track to
double before comparing.

+++ tests/ulm/ulm_tracking2d_test.m
@@ -69,7 +69,10 @@ classdef ulm_tracking2d_test < matlab.unittest.TestCase
             Tracks = ulm.tracking2D(MatTracking, ULM, 'velocityinterp');
 
             testCase.verifyNumElements(Tracks, 1);
-            track = Tracks{1};
+            % tracking2D returns tracks as single; compare as double so
+            % verifyEqual's AbsTol check isn't tripped up by a class
+            % mismatch (single vs double) on otherwise-equal values.
+            track = double(Tracks{1});
             testCase.verifySize(track, [size(track, 1), 5]); % z x vz vx t
 
             % Constant axial speed of 1 px/frame => 1 px/s given scale(3)=1.
